requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-generativeai>=0.8.0
Pillow>=10.0.0
